+++ src/main/java/org/openlmis/referencedata/validate/ProcessingPeriodValidator.java
@@ -28,6 +28,7 @@ import org.springframework.validation.Errors;
 public class ProcessingPeriodValidator implements BaseValidator {
   private static final String START_DATE = "startDate";
   private static final String END_DATE = "endDate";
+  private static final String PROCESSING_SCHEDULE = "processingSchedule";
 
   @Autowired
   private ProcessingPeriodRepository processingPeriodRepository;
@@ -39,6 +40,7 @@ public class ProcessingPeriodValidator implements BaseValidator {
 
   @Override
   public void validate(Object obj, Errors err) {
+    rejectIfEmpty(err, PROCESSING_SCHEDULE, ProcessingPeriodMessageKeys.ERROR_SCHEDULE_NULL);
     rejectIfEmpty(err, START_DATE, ProcessingPeriodMessageKeys.ERROR_START_DATE_NULL);
     rejectIfEmpty(err, END_DATE, ProcessingPeriodMessageKeys.ERROR_END_DATE_NULL);